Swap from up-front inlining to simple method specialization
We flatten out the call-graph based on the data we have available from the AST, but generate a separate monomorphic method for each node in the call graph.

This simplifies things while allowing the JVM to inline/optimize the call-graph later as it wishes, resulting in the same final performance. Future downstream optimization phases can also perform inlining later.